Refactor db functions into three files
